rename: LoginBody 공통 컴포넌트 한 폴더로 이동
